split code into packages - add controllers package - add db package - add lib functions
