+++ lib/utils.go
@@ -1,0 +1,16 @@
+package lib
+
+import (
+	"golang.org/x/crypto/bcrypt"
+)
+
+
+func HashPassword(password string) (string, error) {
+    bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
+    return string(bytes), err
+}
+
+func CheckPasswordHash(password, hash string) bool {
+    err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password))
+    return err == nil
+}
